+++ DawnLib.Dusk/src/API/Definitions/EntityReplacement/EntityReplacementConfig.cs
@@ -16,6 +16,8 @@ public class EntityReplacementConfig(ConfigContext section, string EntityNameRef
         MoonSpawnWeights,
         InteriorSpawnWeights,
         WeatherSpawnWeights,
+        RouteSpawnWeights,
+
         DisableDateCheck
     ];
 }