+++ DawnLib.Dusk/src/API/Definitions/Items/ItemConfig.cs
@@ -7,10 +7,13 @@ namespace Dusk;
 public class ItemConfig(ConfigContext section, string EntityNameReference) : DuskBaseConfig(section, EntityNameReference)
 {
     public ConfigEntry<int>? Cost;
+
     public ConfigEntry<bool>? IsScrapItem;
     public ConfigEntry<bool>? IsShopItem;
+
     public ConfigEntry<bool>? DisableUnlockRequirements = null;
     public ConfigEntry<bool>? DisablePricingStrategy = null;
+
     public ConfigEntry<string>? MoonSpawnWeights;
     public ConfigEntry<string>? InteriorSpawnWeights;
     public ConfigEntry<string>? WeatherSpawnWeights;
@@ -20,13 +23,18 @@ public class ItemConfig(ConfigContext section, string EntityNameReference) : Dus
 
     override internal List<ConfigEntryBase?> _configEntries => [
         Cost,
+
         IsScrapItem,
         IsShopItem,
+
         DisableUnlockRequirements,
         DisablePricingStrategy,
+
         MoonSpawnWeights,
         InteriorSpawnWeights,
         WeatherSpawnWeights,
+        RouteSpawnWeights,
+
         Worth
     ];
 }