+++ DawnLib.Dusk/src/API/Definitions/MapObjects/MapObjectConfig.cs
@@ -8,9 +8,11 @@ public class MapObjectConfig(ConfigContext section, string EntityNameReference) 
     public ConfigEntry<string>? InsideMoonCurveSpawnWeights;
     public ConfigEntry<string>? InsideInteriorCurveSpawnWeights;
     public ConfigEntry<bool>? InsideHazard;
+
     public ConfigEntry<string>? OutsideMoonCurveSpawnWeights;
     public ConfigEntry<string>? OutsideInteriorCurveSpawnWeights;
     public ConfigEntry<bool>? OutsideHazard;
+
     public ConfigEntry<bool>? InsidePrioritiseMoon;
     public ConfigEntry<bool>? OutsidePrioritiseMoon;
 
@@ -18,9 +20,11 @@ public class MapObjectConfig(ConfigContext section, string EntityNameReference) 
         InsideMoonCurveSpawnWeights,
         InsideInteriorCurveSpawnWeights,
         InsideHazard,
+
         OutsideMoonCurveSpawnWeights,
         OutsideInteriorCurveSpawnWeights,
         OutsideHazard,
+
         InsidePrioritiseMoon,
         OutsidePrioritiseMoon
     ];